+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - deleted 'uro'
 - deleted `gmc`
 - added `toBitcoinJS()` for [`bitcoinjs-lib`](https://github.com/bitcoinjs/bitcoinjs-lib) support
-- 
+- added `toBitcore()` for Bitpay's bitcore suppport
 
 0.6.0 / 2015-01-07
 ------------------